+++ Plugin/CleverTapUnity/iOS/CleverTapUnityManager.h
@@ -96,6 +96,7 @@
 - (void)deleteInboxMessageForID:(NSString *)messageId;
 - (void)deleteInboxMessagesForIDs:(NSArray *)messageIds;
 - (void)markReadInboxMessageForID:(NSString *)messageId;
+- (void)markReadInboxMessagesForIDs:(NSArray *)messageIds;
 - (void)recordInboxNotificationViewedEventForID:(NSString *)messageId;
 - (void)recordInboxNotificationClickedEventForID:(NSString *)messageId;
 
@@ -131,5 +132,10 @@
 - (void)suspendInAppNotifications;
 - (void)discardInAppNotifications;
 - (void)resumeInAppNotifications;
+
+#pragma mark - Push Primer
+- (void)promptForPushPermission:(BOOL)showFallbackSettings;
+- (void)promptPushPrimer:(NSDictionary *)json;
+- (void)isPushPermissionGranted;
 
 @end