+++ example/CleverTapUnity.cs
@@ -57,6 +57,41 @@ public class CleverTapUnity : MonoBehaviour
         CleverTapBinding.RecordEvent("Test Unity Event");
         //Invoke("LaunchInbox",30.0f);
 
+
+        //Push primer APIs usage
+        //Half-Interstial Local InApp
+        Dictionary<string, object> item = new Dictionary<string, object>();
+        item.Add("inAppType", "half-interstitial");
+        item.Add("titleText", "Get Notified");
+        item.Add("messageText", "Please enable notifications on your device to use Push Notifications.");
+        item.Add("followDeviceOrientation", true);
+        item.Add("positiveBtnText", "Allow");
+        item.Add("negativeBtnText", "Cancel");
+        item.Add("backgroundColor", "#FFFFFF");
+        item.Add("btnBorderColor", "#0000FF");
+        item.Add("titleTextColor", "#0000FF");
+        item.Add("messageTextColor", "#000000");
+        item.Add("btnTextColor", "#FFFFFF");
+        item.Add("btnBackgroundColor", "#0000FF");
+        item.Add("imageUrl", "https://icons.iconarchive.com/icons/treetog/junior/64/camera-icon.png");
+        item.Add("btnBorderRadius", "2");
+        item.Add("fallbackToSettings", true);
+        CleverTapBinding.PromptPushPrimer(item);
+
+        // //Alert Local InApp
+        // Dictionary<string, object> item = new Dictionary<string, object>();
+        // item.Add("inAppType", "half-interstitial");
+        // item.Add("titleText", "Get Notified");
+        // item.Add("messageText", "Please enable notifications on your device to use Push Notifications.");
+        // item.Add("followDeviceOrientation", true);
+        // item.Add("fallbackToSettings", true);
+        // CleverTapBinding.PromptPushPrimer(item);
+
+        /*Prompt to show hard notification permission dialog
+          true - fallbacks to app's notification settings if permission is denied
+          false - does not fallback to app's notification settings if permission is denied
+        */
+        CleverTapBinding.PromptForPushPermission(false);
     }
 
     /* --------------------------------------------------------------------------------
@@ -72,7 +107,8 @@ public class CleverTapUnity : MonoBehaviour
         CleverTap.CleverTapBinding.RegisterPush();
         // set to 0 to remove icon badge
         CleverTap.CleverTapBinding.SetApplicationIconBadgeNumber(0);
-
+        //Will check whether notification permission is granted or not
+        CleverTapBinding.IsPushPermissionGranted();
     }
 
     /* --------------------------------------------------------------------------------
@@ -84,6 +120,7 @@ public class CleverTapUnity : MonoBehaviour
 #if (UNITY_ANDROID && !UNITY_EDITOR)
         CleverTapBinding.Initialize(CLEVERTAP_ACCOUNT_ID, CLEVERTAP_ACCOUNT_TOKEN);
 
+        //For Android 13+ do ensure to grant notification permission to receive push notifications.
         CleverTapBinding.CreateNotificationChannel("YourChannelId", "Your Channel Name", "Your Channel Description", 5, true);
 
         //CleverTapBinding.CreateNotificationChannelWithSound("YourChannelId","Your Channel Name", "Your Channel Description", 5, true, "Your raw sound file");
@@ -101,6 +138,10 @@ public class CleverTapUnity : MonoBehaviour
         //    CleverTapBinding.SuspendInAppNotifications();
         //    CleverTapBinding.DiscardInAppNotifications();
         //    CleverTapBinding.ResumeInAppNotifications();
+
+        //Returns a boolean to indicate whether notification permission is granted or not
+        bool isPushPermissionGranted = CleverTapBinding.IsPushPermissionGranted();
+        Debug.Log("isPushPermissionGranted"+ isPushPermissionGranted);
 #endif
     }
 
@@ -220,6 +261,19 @@ public class CleverTapUnity : MonoBehaviour
                 Debug.Log("unable to parse json");
             }
         }
+
+        //Delete multiple inbox ids
+        string[] arr = {"1608798445_1676289520","1548315289_1676289520","1608798445_1676289268", "1548315289_1676289268"};
+        CleverTapBinding.DeleteInboxMessagesForIDs(arr);
+        //Marks read for multiple inbox ids
+        CleverTapBinding.MarkReadInboxMessagesForIDs(arr);
+
+        //Delete a single inbox id
+        string inboxId = "1608798445_1676289520";
+        CleverTapBinding.DeleteInboxMessageForID(inboxId);
+        //Marks read for a single inbox id
+        CleverTapBinding.MarkReadInboxMessageForID(inboxId);
+
     }
 
     /* --------------------------------------------------------------------------------
@@ -396,11 +450,18 @@ public class CleverTapUnity : MonoBehaviour
     {
         CleverTapBinding.ShowAppInbox(new Dictionary<string, object>());
     }
+
+    // returns the custom data associated with an in-app notification click
+    void CleverTapInAppNotificationShowCallback(string message) {
+        Debug.Log("unity received inapp notification onShow(): " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
+    }
+
     // returns the custom data associated with an in-app notification click
     void CleverTapInAppNotificationDismissedCallback(string message)
     {
         Debug.Log("unity received inapp notification dismissed: " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
     }
+
     //returns data associated with inbox message item click along with page index and button index
     void CleverTapInboxItemClicked(string message)
     {
@@ -417,9 +478,21 @@ public class CleverTapUnity : MonoBehaviour
     {
         CleverTapBinding.ShowAppInbox(new Dictionary<string, object>());
         Debug.Log("unity received inbox initialized");
-        // string[] arr = {"1608798445_1676289520","1548315289_1676289520","1608798445_1676289268", "1548315289_1676289268"};
-        // CleverTapBinding.DeleteInboxMessagesForIDs(arr);
+
     }
+
+    // returns callback to indicate whether notification permission is granted or not
+    void CleverTapPushNotificationPermissionStatus(string message)
+    {
+        //Ensure to create call the `CreateNotificationChannel` once notification permission is granted to register for receiving push notifications for Android 13+ devices.
+        Debug.Log("CleverTap application isPushPermissionGranted " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
+    }
+
+    // returns the status of push permission response after it's granted/denied
+    void CleverTapOnPushPermissionResponseCallback(string message) {
+        Debug.Log("unity received push permission response: " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
+    }
+    
 
     void CleverTapInboxMessagesDidUpdateCallback()
     {